dex-scalping-monitor/requirements.txt
# DEX Scalping Monitor - Requirements

# Web3 and Blockchain
web3>=6.0.0
eth-account>=0.8.0

# API Requests
requests>=2.31.0
aiohttp>=3.9.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Environment Variables
python-dotenv>=1.0.0

# Notifications (optional - for alerts)
# python-telegram-bot>=20.0  # Uncomment if using Telegram alerts

# Logging and Utils
colorama>=0.4.6
tabulate>=0.9.0

# Async Support
asyncio-mqtt>=0.16.0
